Add scoping function for complete redesign
Stumbled upon the concept of the scoping function during research for cleaner solutions on how to prevent console cheating. A quick test implementation showed that nothing within the scoping function is accessible through console. That also means various functions are now a lot more complicated than they needed to be and can be refactored.